:recycle: Refactor views to lifecycle
