+++ app/src/main/java/hanalyst/application/hanalystclub/lifecycle/Login.java
@@ -1,10 +1,12 @@
-package hanalyst.application.hanalystclub;
+package hanalyst.application.hanalystclub.lifecycle;
 
 import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import hanalyst.application.hanalystclub.R;
 
 public class Login extends Activity {
     Button loginBtn;
