+++ app/src/main/res/layout/activity_analysis.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Analysis">
+    tools:context=".lifecycle.Analysis">
 
     <GridView
         android:id="@+id/grid_attack"
