+++ app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".Login">
+    tools:context=".lifecycle.Login">
 
     <LinearLayout
         android:layout_width="match_parent"
